Revert "[explorer] address page pagination (#12492)"
This reverts commit 67e581a5a1812a7c0806cd9f8f50621cba004fc9.
