--- .changeset/dirty-trainers-thank.md
@@ -1,5 +1,0 @@
----
-'@mysten/sui.js': minor
----
-
-Added FromOrToAddress Transaction Filter
