+++ sdk/typescript/src/types/transactions.ts
@@ -264,7 +264,6 @@ export type SuiTransactionBlockResponseQuery = {
 };
 
 export type TransactionFilter =
-	| { FromOrToAddress: { addr: string } }
 	| { Checkpoint: string }
 	| { FromAndToAddress: { from: string; to: string } }
 	| { TransactionKind: string }
